requirements.txt
asgiref==3.4.1
Django==3.2.13
django-durationwidget==1.0.5
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
